pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "webly"
version = "0.1.0"
description = "Turn any website into a searchable RAG knowledge base"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "Yassin Ali", email = "yelsayed003@gmail.com" }]
keywords = ["rag", "website", "crawling", "embeddings", "vector-search", "llm", "chatbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]

# Core deps — enough to run pipelines headlessly (no UI, no HuggingFace models)
dependencies = [
    "openai>=1.84.0",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "beautifulsoup4>=4.13.4",
    "trafilatura>=2.0.0",
    "lxml>=5.4.0",
    "lxml_html_clean>=0.4.2",
    "faiss-cpu>=1.11.0",
    "numpy>=1.26.3,<2.0",
    "scikit-learn>=1.5.0",
    "scipy>=1.13.0",
    "tiktoken>=0.9.0",
    "regex>=2024.5.15",
]

[project.optional-dependencies]
# Local HuggingFace embedding models
hf = [
    "sentence-transformers>=3.0.1",
    "torch>=2.2.2",
    "torchvision>=0.17.2",
    "torchaudio>=2.2.2",
    "transformers>=4.44.2",
    "huggingface-hub>=0.23.4",
    "tokenizers>=0.19.1",
    "safetensors>=0.4.3",
]
# Streamlit web UI
ui = [
    "streamlit>=1.45.1",
]
# Development tools
dev = [
    "pytest>=8.3.2",
    "ruff>=0.6.9",
]
# Everything
all = [
    "sentence-transformers>=3.0.1",
    "torch>=2.2.2",
    "torchvision>=0.17.2",
    "torchaudio>=2.2.2",
    "transformers>=4.44.2",
    "huggingface-hub>=0.23.4",
    "tokenizers>=0.19.1",
    "safetensors>=0.4.3",
    "streamlit>=1.45.1",
]

[project.urls]
Homepage = "https://github.com/Y-Elsayed/webly"
Repository = "https://github.com/Y-Elsayed/webly"
Issues = "https://github.com/Y-Elsayed/webly/issues"
Changelog = "https://github.com/Y-Elsayed/webly/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "docs*", ".claude*", "webcreeper*"]

[tool.setuptools.package-data]
"webly.chatbot" = ["prompts/*.txt"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
